+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This changelog track changes to the qollage project starting at version v0.1.0
 
+## 0.8.0
+
+* Updated to pyo3 0.25, ureq 3.0, typst 0.13 and quill 0.7.
+* Added explicit link to documentation in the readme.
+* Improved tests' robustness.
+
 ## 0.7.0
 
 * Updated to pyo3 0.24.